+++ lib/instacart_api/client/response_error.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module InstacartApi
-  class Client
-    class ResponseError < StandardError; end
-  end
-end