+++ spec/spec_helper.rb
@@ -12,7 +12,9 @@ require "instacart_api"
 require "webmock/rspec"
 
 root = File.dirname __dir__
-Dir[File.join(root, "spec", "support", "**", "*.rb")].sort.each { |f| require f }
+Dir[File.join(root, "spec", "support", "**", "*.rb")].sort.each do |f|
+  require f
+end
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure